Move editor stuff to own folder
