+++ src/renderer/editor/Editor.tsx
@@ -4,19 +4,19 @@ import 'ace-builds/src-noconflict/mode-python';
 import 'ace-builds/src-noconflict/snippets/python';
 import 'ace-builds/src-noconflict/ext-language_tools';
 import 'ace-builds/src-noconflict/ext-searchbox';
-import uploadSvg from '../../assets/upload.svg';
-import downloadSvg from '../../assets/download.svg';
-import openSvg from '../../assets/open.svg';
-import saveSvg from '../../assets/save.svg';
-import saveAsSvg from '../../assets/save-as.svg';
-import newFileSvg from '../../assets/new-file.svg';
-import consoleSvg from '../../assets/console.svg';
-import consoleClearSvg from '../../assets/console-clear.svg';
-import zoomInSvg from '../../assets/zoom-in.svg';
-import zoomOutSvg from '../../assets/zoom-out.svg';
-import startRobot from '../../assets/start-robot.svg';
-import stopRobot from '../../assets/stop-robot.svg';
-import keyboardKeySvg from '../../assets/keyboard-key.svg';
+import uploadSvg from '../../../assets/upload.svg';
+import downloadSvg from '../../../assets/download.svg';
+import openSvg from '../../../assets/open.svg';
+import saveSvg from '../../../assets/save.svg';
+import saveAsSvg from '../../../assets/save-as.svg';
+import newFileSvg from '../../../assets/new-file.svg';
+import consoleSvg from '../../../assets/console.svg';
+import consoleClearSvg from '../../../assets/console-clear.svg';
+import zoomInSvg from '../../../assets/zoom-in.svg';
+import zoomOutSvg from '../../../assets/zoom-out.svg';
+import startRobot from '../../../assets/start-robot.svg';
+import stopRobot from '../../../assets/stop-robot.svg';
+import keyboardKeySvg from '../../../assets/keyboard-key.svg';
 import addEditorAutocomplete from './addEditorAutocomplete';
 import addEditorTooltips from './addEditorTooltips';
 import './Editor.css';
