+++ src/renderer/editor/addEditorTooltips.tsx
